requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.2
distlib==0.3.8
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.13.1
flake8==7.0.0
flake8-broken-line==1.0.0
flake8-isort==6.1.1
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
google-api-core==2.17.0
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
identify==2.5.33
idna==3.6
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
pathspec==0.12.1
pep8-naming==0.13.3
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.0
prompt-toolkit==3.0.43
protobuf==4.25.2
psycopg-binary==3.1.17
psycopg-pool==3.2.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
pyflakes==3.2.0
pyparsing==3.1.1
pytest==7.4.4
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
SQLAlchemy-Utils==0.41.1
starlette==0.35.1
tomli==2.0.1
types-pyOpenSSL==24.0.0.20240130
types-redis==4.6.0.20240106
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.25.0
vine==5.1.0
virtualenv==20.25.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
